requirements.txt
Flask
python-dotenv
google-generativeai
# Add other dependencies here as you identify them:
# e.g., requests, SpeechRecognition, gTTS, TextBlob, nltk, pytest
